Avoid using doubles (#17)

+++ application/src/main/java/com/mercadolibre/planning/model/me/gateways/planningmodel/dtos/AreaDistribution.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Builder
 public class AreaDistribution {
     private String areaId;
-    private double quantity;
+    private long quantity;
 }
